+++ src/util/cursor-get-text.ts
@@ -6,9 +6,12 @@ import { EditableDocument } from '../cursor-doc/model';
 
 export type RangeAndText = [[number, number], string];
 
-export function currentTopLevelFunction(doc: EditableDocument): RangeAndText {
-    const defunCursor = doc.getTokenCursor(doc.selection.active);
-    const defunStart = defunCursor.rangeForDefun(doc.selection.active)[0];
+export function currentTopLevelFunction(
+    doc: EditableDocument,
+    active: number = doc.selection.active
+): RangeAndText {
+    const defunCursor = doc.getTokenCursor(active);
+    const defunStart = defunCursor.rangeForDefun(active)[0];
     const cursor = doc.getTokenCursor(defunStart);
     while (cursor.downList()) {
         cursor.forwardWhitespace();